requirements.txt
tensorflow
kagglehub
streamlit
pillow
matplotlib
